Update podspec platform
Added Fonnx to Telosnex. It crashes on startup due to the ORT Extensions
dylib not being available. The example app works but logs a warning
about the extensions library being built for 10.13 with a deployment
target of 10.11.

This probably won't fix Telosnex, but it's worth trying just to get
rid of the warning