Whisper & Silero VAD on web; major Whisper refactor
ONNX Web doesn't support ORT extensions.

So, arbitrary audio bytes can no longer be passed as input to Whisper.

This is for the best: there's a complex web of constraints for streaming audio
that limited things to signed PCM-16 little endian anyway.

This also required new models that don't attempt to decode audio, which
are available on HuggingFace.

This also means ORTExtensions can be removed, as it was only included to
get support for audio decoding.

That should be left to a later commit, that also checks on what exactly
ORTExtensions does (ex. IIRC I saw LLM work for it, it may be best to remove
the assets from the build to save about 10 MB, but leave the code in case
its useful again)